chore(frontend): remove DeploysProgress wrapper, render single DeployProgress
Fixes #820.

DeploysProgress was a list container that rendered multiple DeployProgress
cards from a { [index: number]: DeployProgressState } dictionary. However,
activateRunDeployState in useDeployment clears all deploys before every new
run/compile, so there is never more than one active deploy at a time. The
dictionary, index counter, and wrapper component were unnecessary.

- Delete DeploysProgress.tsx
- Move shared types (DeployProgressState, UploadPipelineResp, RunPipeline)
  into DeployProgress.tsx
- Simplify useDeployment to a single DeployProgressState | null
- Render DeployProgress directly in LeftPanel
- Update import in Commands.ts
